Update java-genomics-io to fix bug in Bed entry 5th column outputting coordinates in scientific notation

+++ src/edu/unc/genomics/nucleosomes/NucleosomeCall.java
@@ -143,7 +143,7 @@ public class NucleosomeCall extends ValuedInterval implements Comparable<Nucleos
 	}
 	
 	public double occupancy() {
-		return value;
+		return value.doubleValue();
 	}
 	
 	public void setOccupancy(double value) {
